refactor(search): move home_sites into ops::entangling
home_sites is pure ArchSpec math; hosting it in the nohome placement
strategy forced SearchEngine (the data layer) to depend upward on a
placement peer. Relocating it to ops::entangling removes the inversion.
